edge_monitor/pyproject.toml
[project]
name = "edge_monitor"
version = "0.1.0"
description = "Python Boilerplate contains all the"
readme = "README.md"
authors = [
  {name = "Bejoy Jose", email = "bejoyjose1993@yahoo.in"}
]
maintainers = [
  {name = "Bejoy Jose", email = "bejoyjose1993@yahoo.in"}
]
classifiers = [
    # TODO
]
license = {text = "MIT"}
dependencies = [
  "typer",
  "psutil (>=5.8.0,<6.0.0)",
  "aiohttp (>=3.8.0,<4.0.0)",
  "paho-mqtt (>=1.6.0,<2.0.0)",
  "nvidia-ml-py3 (>=7.352.0,<8.0.0)",
  "fastapi (>=0.118.2,<0.119.0)",
  "uvicorn (>=0.37.0,<0.38.0)"
]
requires-python = ">= 3.9"

[project.optional-dependencies]
test = [
    "coverage",  # testing
    "pytest",  # testing
    "ruff",  # linting
    "ty", # checking types
    "ipdb", # debugging
]

[project.urls]
bugs = "https://github.com/bejoyjose1993/edge_monitor/issues"
changelog = "https://github.com/bejoyjose1993/edge_monitor/blob/master/changelog.md"
homepage = "https://github.com/bejoyjose1993/edge_monitor"

[project.scripts]
run-all = "run_all:main"


[tool.ty]
# All rules are enabled as "error" by default; no need to specify unless overriding.
# Example override: relax a rule for the entire project (uncomment if needed).
# rules.TY015 = "warn"  # For invalid-argument-type, warn instead of error.

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
]

[tool.uv]
package = true
